+++ R/enrichMKEGG.R
@@ -6,6 +6,22 @@
 ##' @inheritParams enrichKEGG
 ##' @return A \code{enrichResult} instance.
 ##' @export
+##' @examples
+##' \dontrun{
+##'   data(geneList, package='DOSE')
+##'   de <- names(geneList)[1:100]
+##'   mkk <- enrichMKEGG(gene = gene,
+##'       organism = 'hsa',
+##'       pvalueCutoff = 1,
+##'       qvalueCutoff = 1)
+##'   head(mkk) 
+##'   
+##'   kk <- gson_KEGG('hsa', KEGG_Type="MKEGG")
+##'   mkk2 <- enrichMKEGG(de,
+##'                   organism = kk,
+##'                   pvalueCutoff = 1)
+##'   head(mkk2)
+##' }
 enrichMKEGG <- function(gene,
                         organism = 'hsa',
                         keyType = 'kegg',
@@ -16,8 +32,16 @@ enrichMKEGG <- function(gene,
                         maxGSSize = 500,
                         qvalueCutoff = 0.2) {
 
-    species <- organismMapper(organism)
-    KEGG_DATA <- prepare_KEGG(species, "MKEGG", keyType)
+    if (inherits(organism, "character")) {                       
+        KEGG_DATA <- prepare_KEGG(species, "MKEGG", keyType)  
+        species <- organismMapper(organism)
+    } else if (inherits(organism, "GSON")) {
+        KEGG_DATA <- organism
+        species <- KEGG_DATA@species
+    } else {
+        stop("organism should be a species name or a GSON object")
+    }
+
     res <- enricher_internal(gene,
                              pvalueCutoff  = pvalueCutoff,
                              pAdjustMethod = pAdjustMethod,