+++ R/gson.R
@@ -3,12 +3,14 @@
 ##'
 ##' @title gson_KEGG
 ##' @param species species
+##' @param KEGG_Type one of "KEGG" and "MKEGG"
+##' @param keyType one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'.
 ##' @return a 'GSON' object
 ##' @author Guangchuang Yu
 ##' @importFrom gson gson
 ##' @export
-gson_KEGG <- function(species) {
-    x <- download_KEGG(species)
+gson_KEGG <- function(species, KEGG_Type="KEGG", keyType="kegg") {
+    x <- download_KEGG(species, KEGG_Type, keyType)
     gsid2gene <- setNames(x[[1]], c("gsid", "gene"))
     gsid2name <- setNames(x[[2]], c("gsid", "name"))
     y <- readLines("https://rest.kegg.jp/info/hsa")
@@ -24,6 +26,11 @@ gson_KEGG <- function(species) {
 
 
 gson_GO <- function(OrgDb, keytype = 'ENTREZID', ont = "BP") {
+
+    if (is(OrgDb, "character")) {
+        require(OrgDb, character.only = TRUE)
+        OrgDb <- eval(parse(text = OrgDb))
+    }
 
     goterms <- AnnotationDbi::Ontology(GO.db::GOTERM)
     if (ont != "ALL") {
@@ -44,13 +51,14 @@ gson_GO <- function(OrgDb, keytype = 'ENTREZID', ont = "BP") {
     species <- AnnotationDbi::species(OrgDb)
     m <- AnnotationDbi::metadata(OrgDb)
     version <- m$value[m$name == "GOSOURCEDATE"]
-    gsname <- m$value[m$name == 'GOSOURCENAME']
-
+    # gsname <- m$value[m$name == 'GOSOURCENAME']
+    gsname <- paste(m$value[m$name == 'GOSOURCENAME'], ont, sep = ";")
     gson(gsid2gene = gsid2gene, 
         gsid2name = gsid2name,
         species = species,
         gsname = gsname,
         version = version,
-        accessed_date = as.character(Sys.Date())
+        accessed_date = as.character(Sys.Date()),
+        keytype = keytype
     )
 }